--- src/app/(profile)/layout.tsx
@@ -1,15 +1,0 @@
-import { Header } from '@/app/(profile)/_component/header/header'
-import { ReactElement } from 'react'
-
-type Props = {
-  children: ReactElement
-}
-
-export default function ({ children }: Props) {
-  return (
-    <>
-      <Header />
-      {children}
-    </>
-  )
-}
